Marvin-Social/requirements.txt
python-dotenv==1.0.0
supabase==2.3.0
openai==1.12.0
requests==2.31.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
schedule==1.2.1 